+++ tests/configuration.rs
@@ -6,10 +6,9 @@
 
 use counter::Counter;
 use microkelvin::{HostStore, StoreRef};
-use rusk_vm::{Contract, GasMeter, NetworkState, Schedule};
-use std::collections::HashMap;
+use rusk_vm::{Config, Contract, GasMeter, NetworkState, OpCosts};
 
-fn execute_contract_with_schedule(schedule: &Schedule) -> u64 {
+fn execute_contract_with_config(config: &'static Config) -> u64 {
     let counter = Counter::new(99);
 
     let code =
@@ -17,7 +16,7 @@ fn execute_contract_with_schedule(schedule: &Schedule) -> u64 {
 
     let store = StoreRef::new(HostStore::new());
     let contract = Contract::new(&counter, code.to_vec(), &store);
-    let mut network = NetworkState::with_schedule(store, schedule);
+    let mut network = NetworkState::with_config(store, config);
 
     let contract_id = network.deploy(contract).expect("Deploy error");
 
@@ -34,51 +33,46 @@ fn execute_contract_with_schedule(schedule: &Schedule) -> u64 {
     gas.spent()
 }
 
+const DEFAULT_CONFIG: Config = Config::new();
+
+const HIGH_COST_CONFIG: Config = Config {
+    op_costs: OpCosts {
+        bit: 10000,
+        add: 10000,
+        mul: 10000,
+        div: 10000,
+        load: 10000,
+        store: 10000,
+        const_decl: 10000,
+        local: 10000,
+        global: 10000,
+        flow: 10000,
+        integer_comp: 10000,
+        float_comp: 10000,
+        float: 10000,
+        conversion: 10000,
+        float_conversion: 10000,
+        reinterpret: 10000,
+        unreachable: 10000,
+        nop: 10000,
+        current_mem: 10000,
+        grow_mem: 10000,
+    },
+    ..Config::new()
+};
+
 #[test]
 fn change_gas_cost_per_op_with_schedule() {
-    let schedule = Schedule::default();
-
-    assert!(execute_contract_with_schedule(&schedule) < 15000);
-
-    let per_type_op_cost: HashMap<String, u32> = [
-        ("bit", 10000),
-        ("add", 10000),
-        ("mul", 10000),
-        ("div", 10000),
-        ("load", 10000),
-        ("store", 10000),
-        ("const", 10000),
-        ("local", 10000),
-        ("global", 10000),
-        ("flow", 10000),
-        ("integer_comp", 10000),
-        ("float_comp", 10000),
-        ("float", 10000),
-        ("conversion", 10000),
-        ("float_conversion", 10000),
-        ("reinterpret", 10000),
-        ("unreachable", 10000),
-        ("nop", 10000),
-        ("current_mem", 10000),
-        ("grow_mem", 10000),
-    ]
-    .iter()
-    .cloned()
-    .map(|(s, c)| (s.to_string(), c))
-    .collect();
-
-    let high_cost_schedule = Schedule {
-        per_type_op_cost,
-        ..Schedule::with_version(1)
-    };
-    assert!(execute_contract_with_schedule(&high_cost_schedule) > 100_000);
+    assert!(execute_contract_with_config(&DEFAULT_CONFIG) < 15000);
+    assert!(execute_contract_with_config(&HIGH_COST_CONFIG) > 100_000);
 }
+
+const NO_METERING_CONFIG: Config = Config {
+    has_metering: false,
+    ..Config::new()
+};
 
 #[test]
 fn no_gas_consumption_when_metering_is_off() {
-    let no_metering_schedule = Schedule {
-        has_metering: false,
-        ..Schedule::with_version(2)
-    };
-    assert_eq!(execute_contract_with_schedule(&no_metering_schedule), 0);
+    assert_eq!(execute_contract_with_config(&NO_METERING_CONFIG), 0);
 }