+++ tests/gas_usage.rs
@@ -8,11 +8,10 @@ use byteorder::{LittleEndian, WriteBytesExt};
 use counter::Counter;
 use microkelvin::{HostStore, StoreRef};
 use register::*;
-use rusk_vm::{Contract, GasMeter, NetworkState, Schedule};
+use rusk_vm::{Contract, GasMeter, NetworkState};
 use stack::*;
 
 fn execute_counter_contract() -> u64 {
-    let schedule = Schedule::default();
     let counter = Counter::new(99);
 
     let code =
@@ -20,7 +19,7 @@ fn execute_counter_contract() -> u64 {
 
     let store = StoreRef::new(HostStore::new());
     let contract = Contract::new(&counter, code.to_vec(), &store);
-    let mut network = NetworkState::with_schedule(store, &schedule);
+    let mut network = NetworkState::new(store);
 
     let contract_id = network.deploy(contract).expect("Deploy error");
 
@@ -38,15 +37,13 @@ fn execute_counter_contract() -> u64 {
 }
 
 fn execute_stack_single_push_pop_contract() -> u64 {
-    let schedule = Schedule::default();
-
     let code =
         include_bytes!("../target/wasm32-unknown-unknown/release/stack.wasm");
     let stack = Stack::new();
 
     let store = StoreRef::new(HostStore::new());
     let contract = Contract::new(&stack, code.to_vec(), &store);
-    let mut network = NetworkState::with_schedule(store, &schedule);
+    let mut network = NetworkState::new(store);
 
     let contract_id = network.deploy(contract).expect("Deploy error");
 
@@ -64,15 +61,13 @@ fn execute_stack_single_push_pop_contract() -> u64 {
 }
 
 fn execute_stack_multi_push_pop_contract(count: u64) -> u64 {
-    let schedule = Schedule::default();
-
     let code =
         include_bytes!("../target/wasm32-unknown-unknown/release/stack.wasm");
     let stack = Stack::new();
 
     let store = StoreRef::new(HostStore::new());
     let contract = Contract::new(&stack, code.to_vec(), &store);
-    let mut network = NetworkState::with_schedule(store, &schedule);
+    let mut network = NetworkState::new(store);
 
     let contract_id = network.deploy(contract).expect("Deploy error");
 
@@ -90,15 +85,13 @@ fn execute_stack_multi_push_pop_contract(count: u64) -> u64 {
 }
 
 fn execute_multiple_transactions_stack_contract(count: u64) -> u64 {
-    let schedule = Schedule::default();
-
     let code =
         include_bytes!("../target/wasm32-unknown-unknown/release/stack.wasm");
     let stack = Stack::new();
 
     let store = StoreRef::new(HostStore::new());
     let contract = Contract::new(&stack, code.to_vec(), &store);
-    let mut network = NetworkState::with_schedule(store, &schedule);
+    let mut network = NetworkState::new(store);
 
     let contract_id = network.deploy(contract).expect("Deploy error");
 
@@ -114,8 +107,6 @@ fn execute_multiple_transactions_stack_contract(count: u64) -> u64 {
 }
 
 fn execute_multiple_register_contract(count: u64) -> u64 {
-    let schedule = Schedule::default();
-
     let code = include_bytes!(
         "../target/wasm32-unknown-unknown/release/register.wasm"
     );
@@ -123,7 +114,7 @@ fn execute_multiple_register_contract(count: u64) -> u64 {
 
     let store = StoreRef::new(HostStore::new());
     let contract = Contract::new(&register, code.to_vec(), &store);
-    let mut network = NetworkState::with_schedule(store, &schedule);
+    let mut network = NetworkState::new(store);
 
     let contract_id = network.deploy(contract).expect("Deploy error");